pyproject.toml
[tool.poetry]
name = "nepal-entity-service"
version = "0.1.0"
description = "A comprehensive service for managing Nepali public entities"
authors = ["Damodar Dahal <damo94761@gmail.com>"]
readme = "README.md"
packages = [{include = "nes"}]

[tool.poetry.dependencies]
python = "^3.12"
pydantic = "^2.12.3"

# API extras
fastapi = {extras = ["standard"], version = "^0.120.4", optional = true}
uvicorn = {version = "^0.24.0", optional = true}

# Scraping extras
wikipedia = {version = "^1.4.0", optional = true}
google-cloud-aiplatform = {version = "^1.38.0", optional = true}
google-auth = {version = "^2.23.0", optional = true}

[tool.poetry.extras]
api = ["fastapi", "uvicorn"]
scraping = ["wikipedia", "google-cloud-aiplatform", "google-auth"]
all = ["fastapi", "uvicorn", "wikipedia", "google-cloud-aiplatform", "google-auth"]

[tool.poetry.scripts]
nes-api = "nes.cli:main"
nes-dev = "nes.cli:dev"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-asyncio = "^0.21.0"
httpx = "^0.27.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
exclude = '''
/(
    \.venv
  | \.git
  | __pycache__
)/
'''